Factor out some common code in a Hydra.Chain.Direct.Util module.
