+++ hydra-node/test/Hydra/Chain/Direct/TxSpec.hs
@@ -8,34 +8,33 @@ import Hydra.Prelude
 import Test.Hydra.Prelude
 
 import Cardano.Binary (serialize)
+import Cardano.Ledger.Alonzo (TxOut)
 import Cardano.Ledger.Alonzo.Data (Data (Data))
-import Cardano.Ledger.Alonzo.Tx (ValidatedTx (ValidatedTx, wits, body), outputs)
+import Cardano.Ledger.Alonzo.Tx (ValidatedTx (ValidatedTx, body, wits), outputs)
+import Cardano.Ledger.Alonzo.TxBody (TxOut (TxOut))
 import Cardano.Ledger.Alonzo.TxWitness (TxWitness (txdats), nullDats, unTxDats)
+import Cardano.Ledger.Crypto (StandardCrypto)
+import Cardano.Ledger.Mary.Value (AssetName, PolicyID, Value (Value))
 import qualified Data.ByteString.Lazy as LBS
 import qualified Data.Map as Map
-import Hydra.Chain (HeadParameters (..), toOnChainTx, PostChainTx (InitTx))
-import Hydra.Chain.Direct.Tx (initTx, observeTx, constructTx)
+import Hydra.Chain (HeadParameters (..), PostChainTx (InitTx), toOnChainTx)
+import Hydra.Chain.Direct.Tx (constructTx, initTx, observeTx)
+import Hydra.Chain.Direct.Util (Era)
 import Hydra.Contract.ContestationPeriod (contestationPeriodFromDiffTime)
 import Hydra.Contract.Head (State (Initial))
 import Hydra.Contract.Party (partyFromVerKey)
+import Hydra.Ledger.Simple (SimpleTx)
 import Hydra.Party (vkey)
 import Plutus.V1.Ledger.Api (toBuiltinData, toData)
 import Test.Cardano.Ledger.Alonzo.Serialisation.Generators ()
 import Test.QuickCheck (counterexample, (===), (==>))
-import Hydra.Ledger.Simple (SimpleTx)
-import Hydra.Chain.Direct (Era)
-import Cardano.Ledger.Mary.Value (Value(Value), PolicyID, AssetName)
-import Cardano.Ledger.Alonzo (TxOut)
-import Cardano.Ledger.Alonzo.TxBody (TxOut(TxOut))
-import Cardano.Ledger.Crypto (StandardCrypto)
 
 spec :: Spec
 spec =
   parallel $ do
     prop "observeTx . constructTx roundtrip" $ \postTx txIn time ->
-      isImplemented postTx ==> -- TODO(SN): test all constructors
-      observeTx (constructTx txIn postTx) === Just (toOnChainTx @SimpleTx time postTx)
-
+      isImplemented postTx
+        ==> observeTx (constructTx txIn postTx) === Just (toOnChainTx @SimpleTx time postTx) -- TODO(SN): test all constructors
     describe "initTx" $ do
       prop "can construct & serialize unsigned initTx" $ \txIn params ->
         let tx = initTx params txIn
@@ -44,7 +43,6 @@ spec =
          in counterexample ("Tx: " <> show tx) $
               counterexample ("Tx serialized size: " <> show len) $
                 len < 16000 -- TODO(SN): use real max tx size
-
       prop "contains some datums" $ \txIn params ->
         let ValidatedTx{wits} = initTx params txIn
             dats = txdats wits
@@ -65,7 +63,8 @@ spec =
         let ValidatedTx{body} = initTx params txIn
             nfts = foldMap txOutNFT $ outputs body
          in counterexample ("NFTs: " <> show nfts) True
-            -- TODO(SN): re-enable length nfts == length (parties params)
+
+-- TODO(SN): re-enable length nfts == length (parties params)
 
 isImplemented :: PostChainTx tx -> Bool
 isImplemented = \case
@@ -75,7 +74,7 @@ isImplemented = \case
 -- | Extract NFT candidates. any single quantity assets not being ADA is a
 -- candidate.
 txOutNFT :: TxOut Era -> [(PolicyID StandardCrypto, AssetName)]
-txOutNFT (TxOut _ value _ ) =
+txOutNFT (TxOut _ value _) =
   mapMaybe findUnitAssets $ Map.toList assets
  where
   (Value _ assets) = value
@@ -84,4 +83,4 @@ txOutNFT (TxOut _ value _ ) =
     (name, _q) <- find unitQuantity $ Map.toList as
     pure (policy, name)
 
-  unitQuantity (_name,q) = q == 1
+  unitQuantity (_name, q) = q == 1
